refactor, clean-up

+++ samoa-instances/src/main/java/org/apache/samoa/instances/Attribute.java
@@ -259,7 +259,7 @@ public class Attribute implements Serializable {
     if (val == null) {
       return -1;
     } else {
-      return val.intValue();
+      return val;
     }
   }
